Apply a function to a faceted piece of data.

Apply a function to a faceted piece of data.

split_apply(data, x, y, FUN, SIMPLIFY = TRUE)

Arguments

data

A data frame.

x

The independent variable (the one which FUN will be applied to).

y

The dependent variable (the one which will facet or subset data).

FUN

A function to apply to the faceted piece of data.

Examples

# Gives the mean \code{wt} when \code{mtcars} has been faceted by \code{cyl}.
facet_fun(mtcars, "wt", "cyl", mean)

#> Warning: argument is not numeric or logical: returning NA

#> Warning: argument is not numeric or logical: returning NA

#> Warning: argument is not numeric or logical: returning NA

#>       4.mpg       4.cyl      4.disp        4.hp      4.drat        4.wt 
#>  26.6636364   4.0000000 105.1363636  82.6363636   4.0709091   2.2857273 
#>      4.qsec        4.vs        4.am      4.gear      4.carb      4.char 
#>  19.1372727   0.9090909   0.7272727   4.0909091   1.5454545          NA 
#>       6.mpg       6.cyl      6.disp        6.hp      6.drat        6.wt 
#>  19.7428571   6.0000000 183.3142857 122.2857143   3.5857143   3.1171429 
#>      6.qsec        6.vs        6.am      6.gear      6.carb      6.char 
#>  17.9771429   0.5714286   0.4285714   3.8571429   3.4285714          NA 
#>       8.mpg       8.cyl      8.disp        8.hp      8.drat        8.wt 
#>  15.1000000   8.0000000 353.1000000 209.2142857   3.2292857   3.9992143 
#>      8.qsec        8.vs        8.am      8.gear      8.carb      8.char 
#>  16.7721429   0.0000000   0.1428571   3.2857143   3.5000000          NA